+++ 1.21/Common/src/main/java/fuzs/puzzleslib/api/client/event/v1/renderer/RenderLivingEvents.java
@@ -13,7 +13,7 @@ public final class RenderLivingEvents {
     public static final EventInvoker<After> AFTER = EventInvoker.lookup(After.class);
 
     private RenderLivingEvents() {
-
+        // NO-OP
     }
 
     @FunctionalInterface