+++ 1.21/Fabric/src/main/java/fuzs/puzzleslib/fabric/api/client/event/v1/FabricClientLevelEvents.java
@@ -15,6 +15,6 @@ public final class FabricClientLevelEvents {
     public static final Event<ClientLevelEvents.Unload> UNLOAD_LEVEL = FabricEventFactory.create(ClientLevelEvents.Unload.class);
 
     private FabricClientLevelEvents() {
-
+        // NO-OP
     }
 }