+++ 1.21/Fabric/src/main/java/fuzs/puzzleslib/fabric/api/client/event/v1/FabricClientPlayerEvents.java
@@ -39,6 +39,6 @@ public final class FabricClientPlayerEvents {
             MovementInputUpdateCallback.class);
 
     private FabricClientPlayerEvents() {
-
+        // NO-OP
     }
 }