+++ 1.21/Fabric/src/main/java/fuzs/puzzleslib/fabric/api/client/event/v1/FabricRendererEvents.java
@@ -84,6 +84,6 @@ public final class FabricRendererEvents {
             ComputeFieldOfViewCallback.class);
 
     private FabricRendererEvents() {
-
+        // NO-OP
     }
 }