+++ 1.21/Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/client/init/FabricItemDisplayOverrides.java
@@ -1,52 +1,46 @@
 package fuzs.puzzleslib.fabric.impl.client.init;
 
 import com.google.common.collect.Maps;
-import fuzs.puzzleslib.fabric.api.core.v1.resources.FabricReloadListener;
-import fuzs.puzzleslib.impl.PuzzlesLibMod;
+import fuzs.puzzleslib.fabric.api.client.event.v1.FabricClientEvents;
 import fuzs.puzzleslib.impl.client.init.ItemDisplayOverridesImpl;
-import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
-import net.fabricmc.fabric.api.resource.ResourceReloadListenerKeys;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.model.BakedModel;
+import net.minecraft.client.resources.model.ModelBakery;
 import net.minecraft.client.resources.model.ModelManager;
 import net.minecraft.client.resources.model.ModelResourceLocation;
-import net.minecraft.server.packs.PackType;
-import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.item.ItemDisplayContext;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.IdentityHashMap;
 import java.util.Map;
 import java.util.Objects;
+import java.util.function.Supplier;
 
 public final class FabricItemDisplayOverrides extends ItemDisplayOverridesImpl {
     @Nullable
     private Map<BakedModel, Map<ItemDisplayContext, BakedModel>> overrideModels;
 
     {
-        ResourceManagerHelper.get(PackType.CLIENT_RESOURCES)
-                .registerReloadListener(new FabricReloadListener(PuzzlesLibMod.id("item_display_overrides"),
-                        (ResourceManager resourceManager) -> {
-                            ModelManager modelManager = Minecraft.getInstance().getModelManager();
-                            this.overrideModels = Maps.newIdentityHashMap();
-                            for (Map.Entry<ModelResourceLocation, Map<ItemDisplayContext, ModelResourceLocation>> overrideEntry : this.overrideLocations.entrySet()) {
-                                BakedModel itemModel = modelManager.getModel(overrideEntry.getKey());
-                                Objects.requireNonNull(itemModel, "item model is null");
-                                this.overrideModels.put(itemModel, overrideEntry.getValue()
-                                        .entrySet()
-                                        .stream()
-                                        .collect(Maps.toImmutableEnumMap(Map.Entry::getKey,
-                                                (Map.Entry<ItemDisplayContext, ModelResourceLocation> entry) -> {
-                                                    BakedModel itemModelOverride = modelManager.getModel(
-                                                            entry.getValue());
-                                                    Objects.requireNonNull(itemModelOverride,
-                                                            "item model override is null"
-                                                    );
-                                                    return itemModelOverride;
-                                                }
-                                        )));
-                            }
-                        }, ResourceReloadListenerKeys.MODELS
-                ));
+        FabricClientEvents.COMPLETE_MODEL_LOADING.register(
+                (Supplier<ModelManager> modelManager, Supplier<ModelBakery> modelBakery) -> {
+                    this.overrideModels = new IdentityHashMap<>();
+                    for (Map.Entry<ModelResourceLocation, Map<ItemDisplayContext, ModelResourceLocation>> overrideEntry : this.overrideLocations.entrySet()) {
+                        BakedModel itemModel = modelManager.get().getModel(overrideEntry.getKey());
+                        Objects.requireNonNull(itemModel, "item model is null");
+                        this.overrideModels.put(itemModel, overrideEntry.getValue()
+                                .entrySet()
+                                .stream()
+                                .collect(Maps.toImmutableEnumMap(Map.Entry::getKey,
+                                        (Map.Entry<ItemDisplayContext, ModelResourceLocation> entry) -> {
+                                            BakedModel itemModelOverride = modelManager.get().getModel(
+                                                    entry.getValue());
+                                            Objects.requireNonNull(itemModelOverride,
+                                                    "item model override is null"
+                                            );
+                                            return itemModelOverride;
+                                        }
+                                )));
+                    }
+        });
     }
 
     public BakedModel getItemModelDisplayOverride(BakedModel itemModel, ItemDisplayContext itemDisplayContext) {