+++ 1.21/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/client/event/NeoForgeClientEventInvokers.java
@@ -168,6 +168,9 @@ public final class NeoForgeClientEventInvokers {
         INSTANCE.register(ModelEvents.AfterModelLoading.class, ModelEvent.BakingCompleted.class, (ModelEvents.AfterModelLoading callback, ModelEvent.BakingCompleted evt) -> {
             callback.onAfterModelLoading(evt::getModelManager);
         });
+        INSTANCE.register(ModelEvents.CompleteModelLoading.class, ModelEvent.BakingCompleted.class, (ModelEvents.CompleteModelLoading callback, ModelEvent.BakingCompleted evt) -> {
+            callback.onCompleteModelLoading(evt::getModelManager, evt::getModelBakery);
+        });
     }
 
     public static void registerEventHandlers() {