+++ 1.21/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.24-1.21] - 2024-08-26
+### Fixed
+- Fix item model display overrides being able to compute before models have been reloaded on Fabric again
+
 ## [v21.0.23-1.21] - 2024-08-23
 ### Changed
 - Enable accessing `HolderLookup$Provider` in `AbstractLootProvider` implementations